+++ custom/motorcontrollers/BrakeableMotorController.java
@@ -1,7 +1,7 @@
 
 // THIS FILE IS TESTED post wpilibj2
 
-package org.usfirst.frc4904.standard.subsystems.motor;
+package org.usfirst.frc4904.standard.custom.motorcontrollers;
 
 public interface BrakeableMotorController extends MotorController {
 	public abstract BrakeableMotorController setBrakeOnNeutral();
