+++ custom/motorcontrollers/CANTalonSRX.java
@@ -2,8 +2,6 @@
 
 package org.usfirst.frc4904.standard.custom.motorcontrollers;
 
-import org.usfirst.frc4904.standard.subsystems.motor.TalonMotorController;
-
 import com.ctre.phoenix.motorcontrol.InvertType;
 import com.ctre.phoenix.motorcontrol.NeutralMode;
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
