+++ custom/motorcontrollers/MotorController.java
@@ -1,5 +1,5 @@
 
-package org.usfirst.frc4904.standard.subsystems.motor;
+package org.usfirst.frc4904.standard.custom.motorcontrollers;
 
 /**
  * An abstract motor interface that replaces the WPILib motor interface
