+++ subsystems/motor/MotorSubsystem.java
@@ -1,6 +1,7 @@
 package org.usfirst.frc4904.standard.subsystems.motor;
 
 import org.usfirst.frc4904.standard.LogKitten;
+import org.usfirst.frc4904.standard.custom.motorcontrollers.MotorController;
 import org.usfirst.frc4904.standard.subsystems.motor.speedmodifiers.IdentityModifier;
 import org.usfirst.frc4904.standard.subsystems.motor.speedmodifiers.SpeedModifier;
